backend/services/summoner_watcher/requirements.txt
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
dnspython==2.4.2
frozenlist==1.4.0
hiredis==2.2.3
idna==3.4
lor-deckcodes==5.0.0
motor==3.3.1
multidict==6.0.4
pymongo==4.5.0
pyot==6.0.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.8.0
yarl==1.9.2
